Initialize my first project and html foundations
Add my initial ideas for how the webpage should be layed out.
Continue my practice writing good commit messages and committing early
and often. Talking like a robot is harder than I thought it would be
and this sentence is proof that I have some work to do when it comes
to writing commit messages with the correct voice.